requirements.txt
Flask>=2.2.3
textblob
requests
python-dotenv
